+++ NEWS.md
@@ -1,2 +1,6 @@
 # MetDNA2InSilicoTool 0.1.0
 - initial commit. An R package to connect KGMN result with other in-silico MS/MS tools
+
+# MetDNA2InSilicoTool 0.1.1
+- add 46STD_exd compounds
+- add is_execute parameter in the `runMsFinderMatch` function